Move extractFromRequest() to internal
