+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/instrumenter/SpanLinksExtractor.java
@@ -7,8 +7,6 @@ package io.opentelemetry.instrumentation.api.instrumenter;
 
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.context.Context;
-import io.opentelemetry.context.propagation.TextMapGetter;
-import io.opentelemetry.context.propagation.TextMapPropagator;
 
 /** Extractor of span links for a request. */
 @FunctionalInterface
@@ -19,13 +17,4 @@ public interface SpanLinksExtractor<REQUEST> {
    * {@code spanLinks}.
    */
   void extract(SpanLinksBuilder spanLinks, Context parentContext, REQUEST request);
-
-  /**
-   * Returns a new {@link SpanLinksExtractor} that will extract a {@link SpanContext} from the
-   * request using configured {@link TextMapPropagator}.
-   */
-  static <REQUEST> SpanLinksExtractor<REQUEST> extractFromRequest(
-      TextMapPropagator propagator, TextMapGetter<REQUEST> getter) {
-    return new PropagatorBasedSpanLinksExtractor<>(propagator, getter);
-  }
 }
