+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/internal/PropagatorBasedSpanLinksExtractor.java
@@ -3,18 +3,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.api.instrumenter;
+package io.opentelemetry.instrumentation.api.internal;
 
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.propagation.TextMapGetter;
 import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.instrumentation.api.instrumenter.SpanLinksBuilder;
+import io.opentelemetry.instrumentation.api.instrumenter.SpanLinksExtractor;
 
-final class PropagatorBasedSpanLinksExtractor<REQUEST> implements SpanLinksExtractor<REQUEST> {
+public final class PropagatorBasedSpanLinksExtractor<REQUEST>
+    implements SpanLinksExtractor<REQUEST> {
+
   private final TextMapPropagator propagator;
   private final TextMapGetter<REQUEST> getter;
 
-  PropagatorBasedSpanLinksExtractor(TextMapPropagator propagator, TextMapGetter<REQUEST> getter) {
+  public PropagatorBasedSpanLinksExtractor(
+      TextMapPropagator propagator, TextMapGetter<REQUEST> getter) {
     this.propagator = propagator;
     this.getter = getter;
   }
